--- views/welcome.hbs
@@ -1,5 +1,0 @@
- <form action="/:name" method="post">
-  <label for="name">Please Enter Your Name</label>
-  <input type="text" name="name" id="name">
-  <input type="submit">
-</form>
